Extract AssetMetadata
